requirements.txt
# Regression Tracker Web Application Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Background tasks
apscheduler>=3.10.0

# Templates
jinja2>=3.1.0

# HTTP client for Jenkins
requests>=2.31.0

# File upload support
python-multipart>=0.0.6

# Rate limiting
slowapi>=0.1.9

# Caching
fastapi-cache2[redis]>=0.2.1

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0

# Production
gunicorn>=21.2.0
